+++ gen.go
@@ -17,3 +17,7 @@ func (m *MakeSet) SetFile(v *ast.File)         { m.file = v }
 func (m *MakeSet) SetTypeName(v string)        { m.typeName = v }
 func (m *MakeSet) SetOutput(v io.Writer)       { m.output = v }
 
+func (m *MakeFill) SetFileSet(v *token.FileSet) { m.fileSet = v }
+func (m *MakeFill) SetFile(v *ast.File)         { m.file = v }
+func (m *MakeFill) SetTypeNames(v []string)     { m.typeNames = v }
+func (m *MakeFill) SetOutput(v io.Writer)       { m.output = v }